+++ R/test_utils.R
@@ -9,3 +9,31 @@ expect_equal_plain_df <-function(object, expected, ..., info = NULL, label = NUL
     eval(bquote(expect_equal(as_plain_df(.(object)),as_plain_df(.(expected)))))
 }
 
+
+#' @noRd
+expect_equal_but_cnt_sgl <-function(object, expected, ..., info = NULL, label = NULL, expected.label = NULL) {
+  object$.signal <- object$.signal[0,]
+  expected$.signal <- expected$.signal[0,]
+    eval(bquote(expect_equal(.(object), .(expected))))
+}
+
+#' @noRd
+expect_equal_but_sgl <-function(object, expected, ..., info = NULL, label = NULL, expected.label = NULL) {
+  object$.signal <- NULL
+  expected$.signal <- NULL
+    eval(bquote(expect_equal(.(object), .(expected))))
+}
+
+#' @noRd
+expect_equal_but_cnt_sgm <-function(object, expected, ..., info = NULL, label = NULL, expected.label = NULL) {
+  object$.segments <- object$.segments[0,]
+  expected$.segments <- expected$.segments[0,]
+    eval(bquote(expect_equal(.(object), .(expected))))
+}
+
+#' @noRd
+expect_equal_but_sgm <-function(object, expected, ..., info = NULL, label = NULL, expected.label = NULL) {
+  object$.segments <- NULL
+  expected$.segments <- NULL
+    eval(bquote(expect_equal(.(object), .(expected))))
+}